+++ app/lib/view/station_cleaning/billing/performance_billing_config_screen.dart
@@ -134,7 +134,7 @@ class _PerformanceBillingConfigScreenState extends State<PerformanceBillingConfi
 
   double get _annualContractValue => _config?.annualContractValue ?? 0;
 
-  double get _contractDays => (_config?.contractDays ?? 0) > 0 ? _config!.contractDays : 365;
+  double get _contractDays => (_config?.contractDays ?? 0) > 0 ? _config!.contractDays.toDouble() : 365;
 
   double _sqftOf(String uid) {
     for (final a in _areas) {
